+++ models/finetune.py
@@ -58,7 +58,7 @@ class Finetune(CLIPInterface):
         enable_pos_emb: bool = False,
     ):
         super().__init__(
-            clip_model.half(),
+            clip_model,
             config,
             token_ids,
             soft_embeddings=torch.zeros([len(attributes)]),
